# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Olake is fastest open-source tool for replicating Databases to Apache Iceberg or Data Lakehouse.",
        "title": "Olake UI API",
        "contact": {
            "name": "OLake",
            "url": "https://olake.io/contact/",
            "email": "hello@olake.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/api/v1/project/{projectid}/sources": {
            "get": {
                "description": "Retrieve a list of all configured sources within a specific project.",
                "tags": [
                    "Sources"
                ],
                "summary": "List all sources",
                "parameters": [
                    {
                        "type": "string",
                        "description": "project id (default is 123)",
                        "name": "projectid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.SourceDataItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "failed to validate request",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400Response"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Error401Response"
                        }
                    },
                    "500": {
                        "description": "failed to retrieve sources",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new source within a project.",
                "tags": [
                    "Sources"
                ],
                "summary": "Create a new source",
                "parameters": [
                    {
                        "type": "string",
                        "description": "project id (default is 123)",
                        "name": "projectid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "source data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateSourceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CreateSourceRequest"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "failed to validate request",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400Response"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Error401Response"
                        }
                    },
                    "413": {
                        "description": "payload too large",
                        "schema": {
                            "$ref": "#/definitions/dto.Error413Response"
                        }
                    },
                    "500": {
                        "description": "failed to create source",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500Response"
                        }
                    }
                }
            }
        },
        "/api/v1/project/{projectid}/sources/spec": {
            "post": {
                "description": "Retrieve the UI spec for a specific source type/version.",
                "tags": [
                    "Sources"
                ],
                "summary": "Get source UI spec",
                "parameters": [
                    {
                        "type": "string",
                        "description": "project id (default is 123)",
                        "name": "projectid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "spec request data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SpecRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SpecResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "failed to validate request",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400Response"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Error401Response"
                        }
                    },
                    "413": {
                        "description": "payload too large",
                        "schema": {
                            "$ref": "#/definitions/dto.Error413Response"
                        }
                    },
                    "500": {
                        "description": "failed to get spec",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500Response"
                        }
                    }
                }
            }
        },
        "/api/v1/project/{projectid}/sources/streams": {
            "post": {
                "description": "Discover and list available data streams from a source.",
                "tags": [
                    "Sources"
                ],
                "summary": "Get source stream catalog",
                "parameters": [
                    {
                        "type": "string",
                        "description": "project id (default is 123)",
                        "name": "projectid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "streams request data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.StreamsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "failed to validate request",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400Response"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Error401Response"
                        }
                    },
                    "413": {
                        "description": "payload too large",
                        "schema": {
                            "$ref": "#/definitions/dto.Error413Response"
                        }
                    },
                    "500": {
                        "description": "failed to get source catalog",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500Response"
                        }
                    }
                }
            }
        },
        "/api/v1/project/{projectid}/sources/test": {
            "post": {
                "description": "Validate the connection to a source using the provided configuration details.",
                "tags": [
                    "Sources"
                ],
                "summary": "Test source connection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "project id (default is 123)",
                        "name": "projectid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "test connection data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SourceTestConnectionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.TestConnectionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "failed to validate request",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400Response"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Error401Response"
                        }
                    },
                    "413": {
                        "description": "payload too large",
                        "schema": {
                            "$ref": "#/definitions/dto.Error413Response"
                        }
                    },
                    "500": {
                        "description": "failed to test connection",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500Response"
                        }
                    }
                }
            }
        },
        "/api/v1/project/{projectid}/sources/versions": {
            "get": {
                "description": "Retrieve the list of available versions for a specific source connector type.",
                "tags": [
                    "Sources"
                ],
                "summary": "Get available source versions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "project id (default is 123)",
                        "name": "projectid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "source type",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.JSONResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.VersionsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "failed to validate request",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400Response"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Error401Response"
                        }
                    },
                    "500": {
                        "description": "failed to get versions",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateSourceRequest": {
            "type": "object",
            "required": [
                "config",
                "name",
                "type",
                "version"
            ],
            "properties": {
                "config": {
                    "type": "string",
                    "example": "{\"host\":\"localhost\",\"port\":5432,\"database\":\"mydb\",\"user\":\"postgres\",\"password\":\"secret\"}"
                },
                "name": {
                    "type": "string",
                    "example": "my-postgres-source"
                },
                "type": {
                    "type": "string",
                    "example": "postgres"
                },
                "version": {
                    "type": "string",
                    "example": "v0.2.7"
                }
            }
        },
...
        "dto.Error400Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Bad request"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "dto.Error401Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Authentication required"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "dto.Error413Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Payload too large"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "dto.Error500Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Internal server error"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "dto.JSONResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "operation completed successfully"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "dto.SourceDataItem": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "string",
                    "example": "{\"host\":\"localhost\",\"port\":5432}"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "created_by": {
                    "type": "string",
                    "example": "admin"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "jobs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.JobDataItem"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "my-postgres-source"
                },
                "type": {
                    "type": "string",
                    "example": "postgres"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-09T12:00:00Z"
                },
                "updated_by": {
                    "type": "string",
                    "example": "admin"
                },
                "version": {
                    "type": "string",
                    "example": "v0.2.7"
                }
            }
        },
...
        "dto.SpecResponse": {
            "type": "object",
            "properties": {
                "spec": {
                    "type": "object"
                },
                "type": {
                    "type": "string",
                    "example": "postgres"
                },
                "version": {
                    "type": "string",
                    "example": "v0.2.7"
                }
            }
        },
...
    },
...
}